service/requirements.txt
Django~=3.2.16
djangorestframework
psycopg2
celery[redis]==5.2.7
flower==1.2.0